move egrid validation functions to separate file
